# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Minimal, production-ready Go API starter built with Fiber, GORM, JWT, and S3-compatible object storage.",
        "title": "Starter Go API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/ai/categorize": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uses Claude Sonnet 4 to extract amount, merchant, category \u0026 type from free-form text.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "AI: categorize transaction from raw OCR text",
                "parameters": [
                    {
                        "description": "OCR text + optional user categories",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.CategorizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.CategorizeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ai/chat": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Free-form Q\u0026A with optional 30-day transaction context attached.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "AI: conversational finance assistant",
                "parameters": [
                    {
                        "description": "User question",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.ChatResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ai/insights": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Aggregates the user's transactions over a date range and asks Claude for a narrative summary, recommendations and anomalies.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "AI: financial insights for the current user",
                "parameters": [
                    {
                        "description": "Date range (defaults to last 30 days)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.InsightsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.InsightsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ai/scan-receipt": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends a base64-encoded receipt image to Claude vision and extracts structured fields.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "AI: scan receipt image (vision)",
                "parameters": [
                    {
                        "description": "Base64 image (jpeg/png/webp)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.ScanReceiptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.ScanReceiptResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ai/suggest-budget": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uses recent expense history to suggest sensible monthly budget caps for each category.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "AI: recommend a monthly budget per category",
                "parameters": [
                    {
                        "description": "Window in months (default 3) + optional wallet filter",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.SuggestBudgetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.SuggestBudgetResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_ganiramadhan_starter-go_internal_dto.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.PaginationMeta"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "github_com_ganiramadhan_starter-go_internal_dto.BudgetSuggestion": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "Food \u0026 Beverage"
                },
                "limit_amount": {
                    "type": "number",
                    "example": 1500000
                },
                "period": {
                    "type": "string",
                    "example": "monthly"
                },
                "reason": {
                    "type": "string",
                    "example": "Average spend over last 3 months: 1.4M"
                }
            }
        },
        "github_com_ganiramadhan_starter-go_internal_dto.CategorizeRequest": {
            "type": "object",
            "required": [
                "text"
            ],
            "properties": {
                "text": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Starbucks Coffee Rp 87.500"
                },
                "user_categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Food",
                        "Shopping",
                        "Transport"
                    ]
                }
            }
        },
        "github_com_ganiramadhan_starter-go_internal_dto.CategorizeResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 87500
                },
                "category": {
                    "type": "string",
                    "example": "Food \u0026 Beverage"
                },
                "confidence": {
                    "type": "number",
                    "example": 0.95
                },
                "merchant_name": {
                    "type": "string",
                    "example": "Starbucks"
                },
                "needs_review": {
                    "type": "boolean",
                    "example": false
                },
                "raw_response": {
                    "type": "object"
                },
                "type": {
                    "type": "string",
                    "example": "expense"
                }
            }
        },
...
        "github_com_ganiramadhan_starter-go_internal_dto.ChatRequest": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "include_context": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "maxLength": 2000,
                    "minLength": 2,
                    "example": "Berapa pengeluaran terbesar saya bulan ini?"
                }
            }
        },
        "github_com_ganiramadhan_starter-go_internal_dto.ChatResponse": {
            "type": "object",
            "properties": {
                "reply": {
                    "type": "string"
                }
            }
        },
...
        "github_com_ganiramadhan_starter-go_internal_dto.InsightsRequest": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string",
                    "example": "2026-04-01"
                },
                "limit": {
                    "type": "integer",
                    "maximum": 500,
                    "minimum": 1,
                    "example": 100
                },
                "to": {
                    "type": "string",
                    "example": "2026-04-30"
                }
            }
        },
        "github_com_ganiramadhan_starter-go_internal_dto.InsightsResponse": {
            "type": "object",
            "properties": {
                "anomalies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "health_score": {
                    "type": "integer",
                    "example": 72
                },
                "period": {
                    "type": "string",
                    "example": "2026-04-01 to 2026-04-30"
                },
                "raw_response": {
                    "type": "object"
                },
                "recommendations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "summary": {
                    "type": "string",
                    "example": "You spent 4.2M this month, 18% above last month."
                },
                "top_categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Food",
                        "Transport",
                        "Shopping"
                    ]
                },
                "total_expense": {
                    "type": "number",
                    "example": 4250000
                },
                "total_income": {
                    "type": "number",
                    "example": 8000000
                }
            }
        },
...
        "github_com_ganiramadhan_starter-go_internal_dto.PaginationMeta": {
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean",
                    "example": true
                },
                "has_previous": {
                    "type": "boolean",
                    "example": false
                },
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "github_com_ganiramadhan_starter-go_internal_dto.ScanReceiptRequest": {
            "type": "object",
            "required": [
                "image_base64"
            ],
            "properties": {
                "image_base64": {
                    "type": "string",
                    "example": "iVBORw0KGgoAAAANSUhEUgAA..."
                },
                "media_type": {
                    "type": "string",
                    "enum": [
                        "image/jpeg",
                        "image/png",
                        "image/webp"
                    ],
                    "example": "image/jpeg"
                },
                "user_categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_ganiramadhan_starter-go_internal_dto.ScanReceiptResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 87500
                },
                "category": {
                    "type": "string",
                    "example": "Food \u0026 Beverage"
                },
                "confidence": {
                    "type": "number",
                    "example": 0.95
                },
                "currency": {
                    "type": "string",
                    "example": "IDR"
                },
                "date": {
                    "type": "string",
                    "example": "2026-05-12"
                },
                "merchant_name": {
                    "type": "string",
                    "example": "Starbucks Sudirman"
                },
                "needs_review": {
                    "type": "boolean",
                    "example": false
                },
                "ocr_text": {
                    "type": "string"
                },
                "raw_response": {
                    "type": "object"
                },
                "type": {
                    "type": "string",
                    "example": "expense"
                }
            }
        },
        "github_com_ganiramadhan_starter-go_internal_dto.SuggestBudgetRequest": {
            "type": "object",
            "properties": {
                "months": {
                    "type": "integer",
                    "maximum": 12,
                    "minimum": 1,
                    "example": 3
                },
                "wallet_id": {
                    "type": "string"
                }
            }
        },
        "github_com_ganiramadhan_starter-go_internal_dto.SuggestBudgetResponse": {
            "type": "object",
            "properties": {
                "notes": {
                    "type": "string"
                },
                "raw_response": {
                    "type": "object"
                },
                "suggestions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.BudgetSuggestion"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer {token}\" to authenticate.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}